requirements.txt
six
tqdm
pytorch=0.3.1
torchtext>=0.2.1
future
allennlp
sentencepiece
gensim